backend/microservices/booking-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
grpcio-health-checking==1.60.0
python-dateutil==2.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
protobuf==4.25.2
asyncio-contextmanager==1.0.0
pybreaker==1.4.0
pydantic==2.5.3
